Use min and max of the C++ STL

+++ src/Git/GitStatusListCtrl.cpp
@@ -3908,7 +3908,7 @@ void CGitStatusListCtrl::Clear()
 bool CGitStatusListCtrl::CheckMultipleDiffs()
 {
 	UINT selCount = GetSelectedCount();
-	if (selCount > max(3, (DWORD)CRegDWORD(L"Software\\TortoiseGit\\NumDiffWarning", 10)))
+	if (selCount > max((DWORD)3, (DWORD)CRegDWORD(L"Software\\TortoiseGit\\NumDiffWarning", 10)))
 	{
 		CString message;
 		message.Format(IDS_STATUSLIST_WARN_MAXDIFF, selCount);
