+++ src/TortoiseShell/ContextMenu.cpp
@@ -2109,7 +2109,7 @@ bool CShellExt::InsertIgnoreSubmenus(UINT &idCmd, UINT idCmdFirst, HMENU hMenu, 
 		else
 			GetMenuTextFromResource(ShellMenuIgnoreSub);
 		menuiteminfo.dwTypeData = stringtablebuffer;
-		menuiteminfo.cch = (UINT)min(wcslen(menuiteminfo.dwTypeData), UINT_MAX);
+		menuiteminfo.cch = (UINT)min(wcslen(menuiteminfo.dwTypeData), (size_t)UINT_MAX);
 
 		InsertMenuItem((topmenu & MENUIGNORE) ? hMenu : subMenu, (topmenu & MENUIGNORE) ? indexMenu++ : indexSubMenu++, TRUE, &menuiteminfo);
 		if (itemStates & ITEMIS_IGNORED)
